refac: ConfigTypes goes to config.ts
