+++ src/init.ts
@@ -2,8 +2,9 @@ import { join as pathJoin } from 'path';
 import { stringify as yamlStringify } from 'yaml';
 import { writeFileSync } from 'fs';
 import { DEFAULT_CONFIG_FILENAME } from './lib/consts';
+import { RawConfigTypes } from './lib/config';
 import { printInfo } from './lib/print';
-import { CommandOpts, RawConfigTypes } from './lib/types';
+import { CommandOpts } from './lib/types';
 
 const DEFAULT_CONFIG: RawConfigTypes = {
   schema: 'lib/type-defs.graphqls',
