+++ test/lib/gql-compile.test.ts
@@ -1,12 +1,13 @@
 import { deepStrictEqual, ok } from 'assert';
 import { join as pathJoin } from 'path';
+import createExecContext from "../../src/lib/exec-context";
 import { rimraf } from '../../src/lib/file';
 import {
   GqlCodegenContext,
   processGqlCompile,
 } from '../../src/lib/gql-compile';
-import { buildConfig } from '../../src/lib/load-config';
-import { ConfigTypes } from '../../src/lib/types';
+import { buildConfig, ConfigTypes } from '../../src/lib/config';
+import { createHash } from "../../src/lib/hash";
 
 const dtsRelDir = 'node_modules/@types/graphql-let';
 const libRelDir = 'node_modules/graphql-let';
@@ -46,10 +47,12 @@ describe('gql-compile', () => {
       const codegenContext: GqlCodegenContext = [];
       const oldGqlContentHashes = new Set<string>();
       // const skippedContext: GqlCodegenContext = [];
+      const execContext = createExecContext(cwd, config, createHash(JSON.stringify(config)))
 
       await processGqlCompile(
-        cwd,
-        config,
+        execContext,
+        // cwd,
+        // config,
         dtsRelDir,
         pathJoin(libRelDir, '__generated__'),
         sourceRelPath,
